Update scm properties to prevent future breakage
The ssh:// protocol for scm.connection is not well supported by the
Jenkins plugin bill of materials and its plugin compatibility tester.
Switch to https:// protocol so that access through teh scm.connection
does not require an ssh private key.

Use the alternate form of scm.developerConnection ssh URL to silence
warnings from the hpi plugin.  Access through scm.developerConnection
continues to use ssh.  Since the plugin is now using automated releases,
it could be switched to https as well, but that change needs discussion
with the plugin maintainer.