Rely on plugin BOM to declare plugin version dependencies
The plugin bill of materials (BOM) defines the versions of various
plugins used by specific Jenkins versions. The plugin bill of materials
includes many commonly used plugins and simplifies dependency management
for plugins that depend on other plugins.

Remove the version declarations in the plugin pom file and rely on the
versions that are tested together in the plugin bill of materials.

https://www.jenkins.io/doc/developer/plugin-development/dependency-management/
provides more details.

Replaces the following pull requests:

* https://github.com/jenkinsci/multibranch-build-strategy-extension-plugin/pull/37
* https://github.com/jenkinsci/multibranch-build-strategy-extension-plugin/pull/39
* https://github.com/jenkinsci/multibranch-build-strategy-extension-plugin/pull/41
* https://github.com/jenkinsci/multibranch-build-strategy-extension-plugin/pull/42
* https://github.com/jenkinsci/multibranch-build-strategy-extension-plugin/pull/44
* https://github.com/jenkinsci/multibranch-build-strategy-extension-plugin/pull/47
* https://github.com/jenkinsci/multibranch-build-strategy-extension-plugin/pull/48